Move testing code from test package -> rest package
